+++ base/src/main/java/org/aya/resolve/context/BindContext.java
@@ -59,7 +59,7 @@ public record BindContext(
   }
 
   @Override
-  public @Nullable ModuleExport2 getModuleLocalMaybe(@NotNull ModuleName.Qualified modName) {
+  public @Nullable ModuleExport getModuleLocalMaybe(@NotNull ModuleName.Qualified modName) {
     return null;
   }
 }
