+++ base/src/main/java/org/aya/resolve/context/PhysicalModuleContext.java
@@ -4,7 +4,6 @@ package org.aya.resolve.context;
 
 import kala.collection.mutable.MutableHashMap;
 import kala.collection.mutable.MutableMap;
-import org.aya.syntax.concrete.stmt.ModuleName;
 import org.aya.syntax.concrete.stmt.Stmt;
 import org.aya.syntax.ref.AnyDefVar;
 import org.aya.syntax.ref.AnyVar;
@@ -18,9 +17,9 @@ import org.jetbrains.annotations.Nullable;
  */
 public non-sealed class PhysicalModuleContext implements ModuleContext {
   public final @NotNull Context parent;
-  public final @NotNull ModuleExport2 exports = new ModuleExport2();
-  public final @NotNull ModuleSymbol2<AnyVar> symbols = new ModuleSymbol2<>();
-  public final @NotNull MutableMap<String, ModuleExport2> modules = MutableHashMap.create();
+  public final @NotNull ModuleExport exports = new ModuleExport();
+  public final @NotNull ModuleSymbol<AnyVar> symbols = new ModuleSymbol<>();
+  public final @NotNull MutableMap<String, ModuleExport> modules = MutableHashMap.create();
   private final @NotNull ModulePath modulePath;
   @Override public @NotNull ModulePath modulePath() { return modulePath; }
 
@@ -33,7 +32,7 @@ public non-sealed class PhysicalModuleContext implements ModuleContext {
 
   @Override public void importModule(
     @NotNull String modName,
-    @NotNull ModuleExport2 modExport,
+    @NotNull ModuleExport modExport,
     @NotNull Stmt.Accessibility accessibility,
     boolean isDefined,
     @NotNull SourcePos sourcePos
@@ -54,7 +53,7 @@ public non-sealed class PhysicalModuleContext implements ModuleContext {
   }
 
   @Override public @NotNull Context parent() { return parent; }
-  @Override public @NotNull ModuleSymbol2<AnyVar> symbols() { return symbols; }
-  @Override public @NotNull MutableMap<String, ModuleExport2> modules() { return modules; }
-  @Override public @NotNull ModuleExport2 exports() { return exports; }
+  @Override public @NotNull ModuleSymbol<AnyVar> symbols() { return symbols; }
+  @Override public @NotNull MutableMap<String, ModuleExport> modules() { return modules; }
+  @Override public @NotNull ModuleExport exports() { return exports; }
 }
