+++ cli-console/src/main/java/org/aya/cli/repl/jline/AyaCompleters.java
@@ -41,12 +41,13 @@ public interface AyaCompleters {
       var fixed = fixWord(word, line);
       var context = repl.replCompiler.getContext();
       context.modules.view().forEach((mod, contents) -> {
-        var modName = mod.ids().joinToString(Constants.SCOPE_SEPARATOR, "", Constants.SCOPE_SEPARATOR);
-        if (!modName.startsWith(fixed.component1())) return;
-        contents.symbols().keysView()
-          .map(name -> (fixed.component2() ? Constants.SCOPE_SEPARATOR : modName) + name)
-          .map(Candidate::new)
-          .forEach(candidates::add);
+        // TODO
+        // var modName = mod.ids().joinToString(Constants.SCOPE_SEPARATOR, "", Constants.SCOPE_SEPARATOR);
+        // if (!modName.startsWith(fixed.component1())) return;
+        // contents.symbols().keysView()
+        //   .map(name -> (fixed.component2() ? Constants.SCOPE_SEPARATOR : modName) + name)
+        //   .map(Candidate::new)
+        //   .forEach(candidates::add);
       });
       context.symbols().keysView().map(Candidate::new).forEach(candidates::add);
     }
