+++ cli-impl/src/test/resources/shared/src/arith/int/base.aya
@@ -1,5 +1,9 @@
-open import arith::nat::base
-open import arith::bool::base
+import arith::nat::base as nbase
+import arith::bool::base as bbase
+
+open nbase
+open bbase
+
 open import relation::binary::path
 
 open inductive Int
