+++ cli-impl/src/test/resources/shared/src/data/vec/properties.aya
@@ -1,6 +1,8 @@
 open import prelude hiding (++)
-open import data::list::base using (List, :<, length)
-open import data::vec::base using (Vec, :>, [], ++)
+import data::list::base as lbase
+import data::vec::base as vbase
+open lbase using (List, :<, length)
+open vbase using (Vec, :>, [], ++)
 
 variable A : Type
 variable n m o : Nat
